+++ Source/OxyPlot.SkiaSharp.Wpf/OxySKElement.cs
@@ -211,10 +211,9 @@ namespace OxyPlot.SkiaSharp.Wpf
         /// <summary>
         /// Returns a reference to the visual object that hosts the dependency object in the visual tree.
         /// </summary>
-        /// <returns> The host window from the visual tree.</returns>
+        /// <returns> The host visual from the visual tree.</returns>
         private Visual GetAncestorVisualFromVisualTree(DependencyObject startElement)
         {
-
             DependencyObject child = startElement;
             DependencyObject parent = VisualTreeHelper.GetParent(child);
             while (parent != null)
@@ -222,7 +221,6 @@ namespace OxyPlot.SkiaSharp.Wpf
                 child = parent;
                 parent = VisualTreeHelper.GetParent(child);
             }
-
             return child is Visual visualChild ? visualChild : Window.GetWindow(this);
         }